% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{display}
\alias{display}
\title{Basic image display.}
\usage{
display(img, ...)
}
\arguments{
\item{img}{A numeric matrix.}

\item{...}{Arguments passed to \code{\link[fields:image.plot]{fields::image.plot()}}. These arguments should
be fully named.}
}
\description{
Display an image that has been read in by \code{\link[=read_tif]{read_tif()}} as it would look in
'ImageJ'. This function wraps \code{\link[fields:image.plot]{fields::image.plot()}}.
}
\examples{
img <- read_tif(system.file("img", "Rlogo.tif", package = "ijtiff"))
display(img[, , 1, 1])  # red channel
display(img[, , 2, 1])  # green channel
display(img[, , 3, 1])  # blue channel

}
