% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_supported_tags}
\alias{get_supported_tags}
\title{Get supported TIFF tags}
\usage{
get_supported_tags()
}
\value{
A named integer vector of supported TIFF tags
}
\description{
Returns a named integer vector of supported TIFF tags. The names are the human-readable
tag names, and the values are the corresponding tag codes.
}
