% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructors.R
\name{ijtiff_img}
\alias{ijtiff_img}
\title{\code{ijtiff_img} class.}
\usage{
ijtiff_img(img, ...)
}
\arguments{
\item{img}{A 4 dimensional array representing an image, indexed by \code{img[y, x, channel, frame]}. This is consistent with the \code{EBImage} package (\url{https://bioconductor.org/packages/EBImage/}).}

\item{...}{Named arguments which are set as attributes.}
}
\description{
A class for images which are read or to be written by the \code{ijtiff} package.
}
\examples{
img <- array(seq_len(2 ^ 4), dim = rep(2, 4))
ijtiff_img(img, bits_per_sample = 8)
}
