% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ijtiff_img}
\alias{print.ijtiff_img}
\title{Print method for an \code{ijtiff_img}.}
\usage{
\method{print}{ijtiff_img}(x, ...)
}
\arguments{
\item{x}{An object of class \link{ijtiff_img}.}

\item{...}{Not currently used.}
}
\value{
The input (invisibly).
}
\description{
Print method for an \code{ijtiff_img}.
}
