% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ijtiff.R
\docType{package}
\name{ijtiff}
\alias{ijtiff}
\alias{ijtiff-package}
\alias{ijtiff-package}
\title{\code{ijtiff}: TIFF I/O for \emph{ImageJ} users}
\description{
Correctly import TIFF files that were saved from ImageJ and write TIFF files
than can be correctly read by ImageJ. Full support for TIFF files with
float/real-numbered pixels. Also supports text image I/O.
}
