% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_tif}
\alias{write_tif}
\title{Write images in TIFF format}
\usage{
write_tif(img, path, bits_per_sample = "auto", compression = "none")
}
\arguments{
\item{img}{A numeric array, the image to write. It's possible to write
single- and multi-plane, images with 1 (grayscale) or more channels.}

\item{path}{file name or a raw vector}

\item{bits_per_sample}{number of bits per sample (numeric scalar). Supported
values are 8, 16, and 32. The default \code{"auto"} automatically picks the
smallest workable value based on the maximum element in \code{img}. For example,
if the maximum element in \code{img} is 789, then 16-bit will be chosen because
789 is greater than 2 ^ 8 - 1 but less than or equal to 2 ^ 16 - 1.}

\item{compression}{A string, the desired compression algorithm. Must be one
of \code{"LZW"}, \code{"none"}, \code{PackBits}", \code{"RLE"}, \code{"JPEG"}, or \code{"deflate"}.}
}
\value{
The input \code{img} (invisibly).
}
\description{
Writes images into a TIFF file.
}
\details{
\itemize{\item For a single-plane, grayscale image, use a matrix \code{img[y, x]}.
\item For a multi-plane, grayscale image, use a 3-dimensional array \code{img[y, x, plane]}. \item For a multi-channel, single-plane image, use a
4-dimensional array with a redundant 4th slot \code{img[y, x, channel, ]} (see
'Examples' for an example of how to do this). \item For a multi-channel,
multi-plane image, use a 4-dimensional array \code{img[y, x, channel, plane]}.}
}
\examples{

img <- read_tif(system.file("img", "Rlogo.tif", package="ijtiff"))
temp_dir <- tempdir()
write_tif(img, paste0(temp_dir, "/", "Rlogo"))
list.files(temp_dir)

}
\seealso{
\code{\link{read_tif}}
}
\author{
Simon Urbanek wrote most of this code for the 'tiff' package. Rory
Nolan lifted it from there and changed it around a bit for this 'ijtiff'
package. Credit should be directed towards Lord Urbanek.
}
